+++ package.json
@@ -29,7 +29,7 @@
   "types": "build/src/index.d.ts",
   "files": [
     "build/",
-    "abi/",
+    "src/abi/",
     "src/",
     "typechain/"
   ],
@@ -62,14 +62,13 @@
     "test:coverage": "c8 -c .nycrc mocha -r ts-node/register tests/**/*.test.ts",
     "test:wait": "ts-node tests/wait.ts",
     "test": "mocha -r ts-node/register tests/**/*.test.ts",
-    "build": "tsc && cp -f typechain/*.d.ts build/typechain && ncp ./abi ./build/abi",
+    "build": "tsc",
     "lint": "gts lint",
     "lint:fix": "gts fix",
     "watch": "tsc --watch",
     "types:fetch": "cd scripts && ./update-abi.sh && cd ../",
-    "types": "rm typechain/*.ts && find ./abi -name \"*.json\" -print0 | xargs -0 -I {} npx typechain --target ethers-v5 --outDir typechain {}",
+    "types": "rm typechain/*.ts && find ./abi -name \"*.json\" -print0 | xargs -0 -I {} npx typechain --target ethers-v5 --outDir src/typechain {}",
     "clean": "gts clean",
-    "compile": "tsc",
     "docs": "typedoc --out docs --hideInPageTOC true --useCodeBlocks true --includeVersion true  src/index.ts && cd scripts && ./post-process-doc.sh && cd ../"
   },
   "publishConfig": {
