+++ src/provider.ts
@@ -9,9 +9,9 @@ import {
 } from 'ethers';
 import {ExternalProvider} from '@ethersproject/providers';
 import {ConnectionInfo, poll} from '@ethersproject/web';
-import {Ierc20Factory as IERC20Factory} from '../typechain/Ierc20Factory';
-import {IEthTokenFactory as IEthTokenFactory} from '../typechain/IEthTokenFactory';
-import {Il2BridgeFactory as IL2BridgeFactory} from '../typechain/Il2BridgeFactory';
+import {Ierc20Factory as IERC20Factory} from './typechain/Ierc20Factory';
+import {IEthTokenFactory as IEthTokenFactory} from './typechain/IEthTokenFactory';
+import {Il2BridgeFactory as IL2BridgeFactory} from './typechain/Il2BridgeFactory';
 import {
   Address,
   BalancesMap,
