+++ src/typechain/Ierc1271.d.ts
