+++ src/wallet.ts
@@ -23,9 +23,9 @@ import {
 } from './types';
 import {ProgressCallback} from '@ethersproject/json-wallets';
 import {AdapterL1, AdapterL2} from './adapters';
-import {IZkSync} from '../typechain/IZkSync';
-import {Il1Bridge} from '../typechain/Il1Bridge';
-import {Il2Bridge} from '../typechain/Il2Bridge';
+import {IZkSync} from './typechain/IZkSync';
+import {Il1Bridge} from './typechain/Il1Bridge';
+import {Il2Bridge} from './typechain/Il2Bridge';
 
 /**
  * A `Wallet` is an extension of {@link ethers.Wallet} with additional features for interacting with zkSync Era.
