+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "./node_modules/gts/tsconfig-google.json",
   "compilerOptions": {
-    "rootDir": ".",
+    "rootDir": "./src",
     "outDir": "build",
     "module": "CommonJS",
     "target": "ES2019",
@@ -18,6 +18,5 @@
   },
   "include": [
     "src/**/*.ts",
-    "tests/**/*.ts"
   ]
 }
